Move getAllContractFiles to utils
