+++ plugin/src/atoms/index.ts
@@ -6,7 +6,6 @@ export {
   type CompilationKeys,
   type SetCompilationValue
 } from './compilation'
-
 export { contractsAtom, selectedContractAtom } from './compiledContracts'
 export { accountAtom, providerAtom } from './connection'
 export { deployedContractsAtom, deployedSelectedContractAtom } from './deployedContracts'
